[mathml] Make math display values on non-MathML elements compute to flow
This CL aligns with a recent resolution from the CSSWG:
https://github.com/w3c/csswg-drafts/issues/5385#issuecomment-690462275

This also ensures that LayoutObjectFactory::CreateMath always receives
a MathML element.

The new test fails when MathML Core (in particular LayoutNG) is
disabled because math display values are not supported. Subtests fail
when MathML Core is enabled because multiple display values are not
supported.

Bug: 1127222, 6606
Change-Id: Ica5558f221960d0f80609dfe0e56c029de7e9c3a