scrollbar-gutter: "force" keyword
Implementation of "force" for the scrollbar-gutter CSS property.
This keyword forces a gutter on a non-scrollable element, so
it can be accurately aligned with nearby scrollable areas
(e.g. when implementing a header).

css_properties.json5:
scrollbar-gutter becomes non-inheritable, as otherwise the children
inside a "force" element would also get gutters of their own,
messing up the layout. This matches the current editor's draft:
https://drafts.csswg.org/css-overflow-4/#scrollbar-gutter-property

paint_layer:
Assign a scrollable area to "force" elements, in order to calculate
the size of scrollbar gutters.

layout_box:
Assign a layer to "force" elements.
Take "force" into account when checking for scrollbar gutters.
ScrollWidth and ScrollHeight defer to the scrollable area for
"force" elements.

Updated and expanded the scrollbar-gutter tests.

Bug: 710214
Change-Id: Iea25ded4b7b635edeb513a93e30645326eb051e7